refactor(api): PR-6 commit 6 — queue 子包 20 routes 内部拆 3 文件
server.py 2952 → 2382 行（−570），20 queue routes + 4 BaseModel + 11 outputs
helper 抽到 api/routers/queue/ 3 文件子包。最大的单 commit。

  api/routers/queue/__init__.py         11   说明文档
  api/routers/queue/lifecycle.py       260   12 routes: list/enqueue/hold(get/post)/
                                              release/reorder/get/cancel/pause/resume/
                                              retry/delete + is_pausable 注入
  api/routers/queue/io.py               83   3 routes: export/import/snapshot/config
  api/routers/queue/outputs.py         305   5 routes: outputs/outputs.zip/
                                              export-outputs/output/{filename}/open-folder
                                              + 11 _task_output_* helpers
  api/schemas/queue.py                  24   EnqueueRequest / ReorderRequest /
                                              ImportRequest / ExportOutputsBody

3-文件 internal split 按职责切（来自 planning doc）：
  lifecycle：task 状态机（list/enqueue/cancel/pause/resume 等）
  io：数据 import / export + task snapshot config 读取
  outputs：训练产物文件列表 / 下载 / 打包 / 系统级 open-folder

3 sub-router 独立 `router = APIRouter()`，api/app.py 三个 include_router。
**注册顺序**：io 必须在 lifecycle 之前 —— FastAPI 按 path 定义顺序匹配，
`/api/queue/export` / `/api/queue/import` 否则会被 `/api/queue/{task_id}`
的整数解析截胡 422（原 server.py 也是这个顺序）。

test_route_invariants `glob("*.py")` → `rglob("*.py")` 跟进，递归覆盖
queue/ 等子包内的 router 文件。

回归：PR-1 安全网 152 + studio_server 44 + queue_io 7 + ... = **243/243 全绿**。
route snapshot 不变（160 routes）。
