+++ studio/server.py
@@ -2311,609 +2311,16 @@ def version_thumb(
     return _thumb_response(f, size)
 
 
-# ---------------------------------------------------------------------------
-# /api/queue, /api/logs, /api/events  (P3)
-# ---------------------------------------------------------------------------
+# /api/queue/* (20 routes) + EnqueueRequest / ReorderRequest / ImportRequest /
+# ExportOutputsBody + _task_output_* helpers 已 PR-6 commit 6 抽到
+# api/routers/queue/{lifecycle,io,outputs}.py + api/schemas/queue.py（3 文件
+# 子包 internal split：12 lifecycle + 3 io + 5 outputs = 20）。
 
-
-class EnqueueRequest(BaseModel):
-    config_name: str
-    name: Optional[str] = None
-    priority: int = 0
-
-
-class ReorderRequest(BaseModel):
-    ordered_ids: list[int]
-
-
-# `_supervisor()` 已 PR-6 commit 2 抽到 api/deps.py，server.py 内剩余 handler 复用同函数
+# `_supervisor()` / `_resolve_anima_model_paths()` 已 PR-6 commit 2/5 抽到
+# api/deps.py，server.py 内剩余 handler 复用同函数
 from .api.deps import _resolve_anima_model_paths, _supervisor  # noqa: E402
 
 
-# 导入 / 导出必须放在 /api/queue/{task_id} 之前，否则 "export" / "import" 会
-# 被当成 task_id 走整数解析。
-class ImportRequest(BaseModel):
-    payload: dict[str, Any]
-
-
-@app.get("/api/queue/export")
-def export_queue(ids: str = "") -> Response:
-    """`?ids=1,2,3` 指定导出的任务，缺省导出全部。
-
-    响应带 `Content-Disposition: attachment` —— 前端 <a download> 直链就能触发
-    浏览器原生下载（和 train.zip / outputs.zip 一套范式）。导出/失败 publish
-    queue_export_ready / _failed SSE，前端用来清 app-side spinner + 弹 toast。
-    body 仍是合法 JSON，tests / 程序化调用方拿 resp.json() 不受影响。
-    """
-    import json as _json
-
-    if ids.strip():
-        try:
-            id_list = [int(x) for x in ids.split(",") if x.strip()]
-        except ValueError:
-            raise HTTPException(400, "ids must be comma-separated integers")
-    else:
-        with db.connection_for() as conn:
-            id_list = [t["id"] for t in db.list_tasks(conn)]
-    try:
-        payload = queue_io.export_tasks(id_list)
-    except Exception as exc:
-        bus.publish({"type": "queue_export_failed", "error": str(exc)})
-        raise
-
-    body = _json.dumps(payload, ensure_ascii=False, indent=2).encode("utf-8")
-    filename = f"queue_{time.strftime('%Y-%m-%d_%H-%M-%S')}.json"
-    bus.publish({"type": "queue_export_ready"})
-    return Response(
-        content=body,
-        media_type="application/json",
-        headers={"Content-Disposition": f'attachment; filename="{filename}"'},
-    )
-
-
-@app.post("/api/queue/import")
-def import_queue(body: ImportRequest) -> dict[str, Any]:
-    try:
-        return queue_io.import_tasks(body.payload)
-    except (ValueError, presets_io.PresetError) as exc:
-        raise HTTPException(400, str(exc)) from exc
-
-
-@app.get("/api/queue")
-def list_queue(
-    status: Optional[str] = None,
-    include_generate: bool = False,
-) -> dict[str, Any]:
-    """队列默认隐藏 generate 测试出图任务（commit 15 P0-2）。
-
-    generate task 走 daemon 不占 train slot，且生命周期短（出完图就结束），
-    出现在队列里只会让用户混淆"为什么队列卡住"。需要排查时加
-    `?include_generate=true` 兜底。
-    """
-    if status and status not in db.VALID_STATUSES:
-        raise HTTPException(400, f"unknown status: {status}")
-    with db.connection_for() as conn:
-        items = db.list_tasks(conn, status=status)
-    if not include_generate:
-        items = db.filter_out_task_types(items, ("generate", "reg_ai"))
-    # ADR 0006 PR-4 — is_pausable 信号每行注入（§8.1 / 上面 get_queue_item 注释）
-    try:
-        sup = _supervisor()
-        for it in items:
-            it["is_pausable"] = sup.is_task_pausable(int(it["id"]))
-    except HTTPException:
-        for it in items:
-            it["is_pausable"] = False
-    return {"items": items}
-
-
-@app.post("/api/queue")
-def enqueue(body: EnqueueRequest) -> dict[str, Any]:
-    cfg_path = USER_PRESETS_DIR / f"{body.config_name}.yaml"
-    if not cfg_path.exists():
-        raise HTTPException(404, f"preset not found: {body.config_name}")
-    name = body.name or body.config_name
-    with db.connection_for() as conn:
-        task_id = db.create_task(
-            conn, name=name, config_name=body.config_name, priority=body.priority
-        )
-        task = db.get_task(conn, task_id)
-    bus.publish(
-        {"type": "task_state_changed", "task_id": task_id, "status": "pending"}
-    )
-    return task or {"id": task_id}
-
-
-@app.get("/api/queue/hold")
-def get_queue_hold() -> dict[str, Any]:
-    """查看当前队列挂起状态 + 等待恢复调度的 pending task 数（UI banner 用）。"""
-    with db.connection_for() as conn:
-        held = db.get_queue_held(conn)
-        pending = db.list_tasks(conn, status="pending")
-    return {"held": held, "pending_waiting": len(pending)}
-
-
-@app.post("/api/queue/hold")
-def hold_queue() -> dict[str, Any]:
-    """挂起队列：dispatcher 不再拉新 task。已 running 的不受影响（ADR §3.2）。
-
-    "同时暂停 running task" 由前端 modal 拆成两步：先调本 endpoint，再
-    单独调 `/api/queue/{id}/pause`。后端不做合一操作。
-    """
-    with db.connection_for() as conn:
-        db.set_queue_held(conn, True)
-    bus.publish({"type": "queue_hold_changed", "held": True})
-    return {"held": True}
-
-
-@app.post("/api/queue/release")
-def release_queue() -> dict[str, Any]:
-    """恢复调度：dispatcher 重新按 priority + created_at 拉 pending。"""
-    with db.connection_for() as conn:
-        db.set_queue_held(conn, False)
-    bus.publish({"type": "queue_hold_changed", "held": False})
-    return {"held": False}
-
-
-@app.post("/api/queue/reorder")
-def reorder_queue(body: ReorderRequest) -> dict[str, Any]:
-    with db.connection_for() as conn:
-        db.reorder(conn, body.ordered_ids)
-    return {"reordered": len(body.ordered_ids)}
-
-
-@app.get("/api/queue/{task_id}")
-def get_queue_item(task_id: int) -> dict[str, Any]:
-    with db.connection_for() as conn:
-        task = db.get_task(conn, task_id)
-    if not task:
-        raise HTTPException(404)
-    # ADR 0006 PR-4 — is_pausable 信号让 UI 决定是否显示暂停按钮（§8.1）。
-    # 仅 supervisor 跑得起来时计算；空载（test / 启动期）默认 False。
-    try:
-        task["is_pausable"] = _supervisor().is_task_pausable(task_id)
-    except HTTPException:
-        task["is_pausable"] = False
-    return task
-
-
-@app.get("/api/queue/{task_id}/snapshot/config")
-def get_task_snapshot_config(task_id: int) -> dict[str, Any]:
-    """ADR-0007 §11.7：返回 task 启动时冻结的 config。
-
-    返回 ``{"yaml": str, "config": dict}``。task 不存在 / 无 snapshot → 404。
-    UI [关联配置] tab 用此 + 触发 "套用此配置" 路由跳转到 ⑦ 训练 phase + prefill。
-    """
-    from . import task_snapshot
-    with db.connection_for() as conn:
-        task = db.get_task(conn, task_id)
-    if not task:
-        raise HTTPException(404, "task not found")
-    data = task_snapshot.read_snapshot_config(task_id)
-    if data is None:
-        raise HTTPException(404, "snapshot not found")
-    return data
-
-
-@app.post("/api/queue/{task_id}/cancel")
-def cancel_task(task_id: int) -> dict[str, Any]:
-    if not _supervisor().cancel(task_id):
-        # 可能任务已结束 / 不在 supervisor 控制
-        with db.connection_for() as conn:
-            task = db.get_task(conn, task_id)
-        if not task:
-            raise HTTPException(404)
-        if task["status"] in db.TERMINAL_STATUSES:
-            raise HTTPException(400, f"task already {task['status']}")
-        raise HTTPException(409, "cancel rejected (state mismatch)")
-    return {"task_id": task_id, "canceled": True}
-
-
-# ---------------------------------------------------------------------------
-# ADR 0006 — pause / resume / hold endpoints。
-# ---------------------------------------------------------------------------
-
-
-@app.post("/api/queue/{task_id}/pause")
-def pause_task(task_id: int) -> dict[str, Any]:
-    """暂停 running task（ADR §4.1 / §4.3）。
-
-    异步：立即返回；UI 端 modal 订阅 SSE 看保存进度。supervisor 收到子进程
-    `__EVENT__:pause_state` 后把 status 写为 paused 并 publish task_state_changed。
-    """
-    ok, reason = _supervisor().pause(task_id)
-    if not ok:
-        # 区分客户端错误（404/409）vs 状态机不允许（409）
-        with db.connection_for() as conn:
-            task = db.get_task(conn, task_id)
-        if not task:
-            raise HTTPException(404, "task not found")
-        raise HTTPException(409, reason or "pause rejected")
-    return {"task_id": task_id, "pause_pending": True}
-
-
-@app.post("/api/queue/{task_id}/resume")
-def resume_task(task_id: int) -> dict[str, Any]:
-    """恢复 paused task（ADR 0006 §6 路径 A）。
-
-    流程：
-      1. 校验 status == 'paused' + paused_state_path 文件存在
-      2. task → pending（**保留 paused_* 字段**，cmd_builder 下轮 dispatch 读它）
-      3. supervisor 下次 _tick 自然 pick up，cmd 加 `--resume-state <pt>`，
-         bootstrap_phase 读 sibling .config.json snapshot 覆盖 args
-      4. 子进程 load_training_state 成功后 emit `resume_state_loaded` →
-         supervisor `_clear_pause_artifacts` 清文件对 + db 字段
-
-    文件丢失 → 409（ADR §5.5：引导用户走 ResumeFieldPicker 起新 task）。
-    """
-    with db.connection_for() as conn:
-        task = db.get_task(conn, task_id)
-        if not task:
-            raise HTTPException(404, "task not found")
-        if task["status"] != "paused":
-            raise HTTPException(
-                409, f"task status is {task['status']!r}, not paused"
-            )
-        state_path = task.get("paused_state_path")
-        config_path = task.get("paused_config_path")
-        if not state_path or not Path(state_path).exists():
-            raise HTTPException(
-                409,
-                f"paused state file missing: {state_path!r}; "
-                f"use ResumeFieldPicker to start a fresh task from another .pt",
-            )
-        if config_path and not Path(config_path).exists():
-            # snapshot 缺失虽然不致命（bootstrap_phase 会沿用 args.config yaml），
-            # 但 resume 语义会漂；按 ADR §5.7 严格 freeze 原则，拒绝继续。
-            raise HTTPException(
-                409,
-                f"paused config snapshot missing: {config_path!r}; "
-                f"cannot guarantee config freeze, refusing to resume",
-            )
-        db.update_task(
-            conn, task_id,
-            status="pending",
-            started_at=None,
-            finished_at=None,
-            exit_code=None,
-            error_msg=None,
-        )
-    bus.publish({"type": "task_state_changed", "task_id": task_id, "status": "pending"})
-    return {"task_id": task_id, "status": "pending"}
-
-
-@app.post("/api/queue/{task_id}/retry")
-def retry_task(task_id: int) -> dict[str, Any]:
-    """已结束任务重新入队：复制完整训练上下文创建新 task。
-
-    需要复制的字段（PP6.1+ 引入；老的 retry 只复制 name/config_name/priority
-    会让 supervisor 走老降级路径用全局 preset 而不是 version 私有 config，
-    导致重试参数与原任务不一致）：
-    - config_path：version 私有 config 的绝对路径
-    - project_id / version_id：用于 monitor_state_path 解析与 stage 推进
-
-    不复制：status / pid / *_at / exit_code / error_msg / monitor_state_path
-    （都是「上次跑」的产物；新任务从 pending 开始，supervisor 会重新解析）。
-    """
-    with db.connection_for() as conn:
-        original = db.get_task(conn, task_id)
-        if not original:
-            raise HTTPException(404)
-        if original["status"] not in db.TERMINAL_STATUSES:
-            raise HTTPException(400, "only terminal tasks can be retried")
-        new_id = db.create_task(
-            conn,
-            name=original["name"],
-            config_name=original["config_name"],
-            priority=original["priority"],
-        )
-        copy_fields: dict[str, Any] = {}
-        for k in ("config_path", "project_id", "version_id"):
-            if original.get(k) is not None:
-                copy_fields[k] = original[k]
-        if copy_fields:
-            db.update_task(conn, new_id, **copy_fields)
-        new_task = db.get_task(conn, new_id)
-    bus.publish(
-        {"type": "task_state_changed", "task_id": new_id, "status": "pending"}
-    )
-    return new_task or {"id": new_id}
-
-
-# ---------------------------------------------------------------------------
-# /api/queue/{task_id}/outputs — 查看 / 下载训练产物
-# ---------------------------------------------------------------------------
-
-
-_LOCALHOST_HOSTS = {"127.0.0.1", "::1", "localhost"}
-_LORA_EXTS = {".safetensors", ".ckpt", ".pt", ".bin"}
-
-
-def _task_output_dir(task: dict[str, Any]) -> Optional[Path]:
-    """根据 task 推断 output 目录：versions/{label}/output。
-
-    没 project_id / version_id 的老任务（PP1 之前）→ 返回 None；调用方应该
-    处理为「无 output 目录」。
-    """
-    pid = task.get("project_id")
-    vid = task.get("version_id")
-    if not (pid and vid):
-        return None
-    with db.connection_for() as conn:
-        v = versions.get_version(conn, int(vid))
-        p = projects.get_project(conn, int(pid))
-    if not v or not p or v["project_id"] != pid:
-        return None
-    return versions.version_dir(int(pid), p["slug"], v["label"]) / "output"
-
-
-class _ExportOutputsBody(BaseModel):
-    files: Optional[list[str]] = None
-
-
-def _select_task_output_files(task_id: int, files: Optional[list[str]] = None) -> tuple[dict[str, Any], list[Path], bool]:
-    with db.connection_for() as conn:
-        task = db.get_task(conn, task_id)
-    if not task:
-        raise HTTPException(404)
-    out_dir = _task_output_dir(task)
-    if not out_dir or not out_dir.exists():
-        raise HTTPException(404, "no output dir")
-    all_files = _iter_task_output_files(out_dir, task_id)
-    if not all_files:
-        raise HTTPException(404, "empty output dir")
-    if not files:
-        return task, all_files, False
-    by_path = {_task_output_relpath(out_dir, f): f for f in all_files}
-    selected: list[Path] = []
-    missing: list[str] = []
-    for name in files:
-        _safe_output_relpath_or_400(out_dir, name)
-        f = by_path.get(name)
-        if f:
-            selected.append(f)
-        else:
-            missing.append(name)
-    if missing:
-        raise HTTPException(404, f"file(s) not found: {', '.join(missing)}")
-    if not selected:
-        raise HTTPException(400, "empty files list")
-    return task, selected, True
-
-
-def _write_outputs_zip(dest: Path, out_dir: Path, selected: list[Path]) -> None:
-    with zipfile.ZipFile(dest, "w", compression=zipfile.ZIP_STORED, allowZip64=True) as zf:
-        for f in selected:
-            zf.write(f, arcname=_task_output_relpath(out_dir, f))
-
-
-def _task_archive_basename(task: dict[str, Any]) -> Optional[str]:
-    """task 关联 project / version → "{slug}-{label}"，用作 outputs zip 文件名
-    前缀。和 train.zip 命名风格一致（PP7：{slug}-{label}.train.zip）。
-
-    没 project / version → None，调用方 fallback 到 task_{id}。
-    """
-    pid = task.get("project_id")
-    vid = task.get("version_id")
-    if not (pid and vid):
-        return None
-    with db.connection_for() as conn:
-        v = versions.get_version(conn, int(vid))
-        p = projects.get_project(conn, int(pid))
-    if not v or not p or v["project_id"] != pid:
-        return None
-    return f"{p['slug']}-{v['label']}"
-
-
-def _is_loopback(request: Request) -> bool:
-    client = request.client
-    return bool(client and client.host in _LOCALHOST_HOSTS)
-
-
-def _task_output_kind(path: Path) -> str:
-    name = path.name
-    suffix = path.suffix.lower()
-    if name.startswith("training_state_") and suffix == ".pt":
-        return "training_state"
-    if name.startswith("pause_step_") and suffix == ".pt":
-        return "pause_state"
-    if name == "auto_epoch_state.pt":
-        return "auto_epoch_state"
-    if suffix in _LORA_EXTS and not name.startswith("training_state_"):
-        return "lora"
-    return "other"
-
-
-def _task_output_relpath(out_dir: Path, path: Path) -> str:
-    return path.relative_to(out_dir).as_posix()
-
-
-def _iter_task_output_files(out_dir: Path, task_id: int) -> list[Path]:
-    files = [p for p in out_dir.iterdir() if p.is_file()]
-    state_dir = out_dir / "state" / f"task_{task_id}"
-    if state_dir.exists():
-        files.extend(
-            p for p in state_dir.rglob("*")
-            if p.is_file() and _task_output_kind(p) in {"training_state", "pause_state", "auto_epoch_state"}
-        )
-    return sorted(files, key=lambda p: _task_output_relpath(out_dir, p))
-
-
-def _safe_output_relpath_or_400(base: Path, relpath: str) -> Path:
-    if not relpath or relpath.startswith(("/", "\\")):
-        raise HTTPException(400, "invalid path")
-    parts = Path(relpath.replace("\\", "/")).parts
-    if any(part in ("", ".", "..") for part in parts):
-        raise HTTPException(400, "invalid path")
-    return _safe_join_or_400(base, *parts)
-
-
-@app.get("/api/queue/{task_id}/outputs")
-def list_task_outputs(task_id: int, request: Request) -> dict[str, Any]:
-    """列出 task 关联 version 的 output 目录里所有文件。"""
-    with db.connection_for() as conn:
-        task = db.get_task(conn, task_id)
-    if not task:
-        raise HTTPException(404)
-    out_dir = _task_output_dir(task)
-    files: list[dict[str, Any]] = []
-    if out_dir and out_dir.exists():
-        for f in _iter_task_output_files(out_dir, task_id):
-            relpath = _task_output_relpath(out_dir, f)
-            try:
-                st = f.stat()
-            except OSError:
-                continue
-            kind = _task_output_kind(f)
-            files.append({
-                "name": f.name,
-                "path": relpath,
-                "size": st.st_size,
-                "mtime": st.st_mtime,
-                "kind": kind,
-                "is_lora": kind == "lora",
-            })
-    return {
-        "task_id": task_id,
-        "output_dir": str(out_dir) if out_dir else None,
-        "exists": bool(out_dir and out_dir.exists()),
-        "supports_open_folder": _is_loopback(request),
-        "files": files,
-        "archive_basename": _task_archive_basename(task),
-    }
-
-
-@app.get("/api/queue/{task_id}/outputs.zip")
-def download_task_outputs_zip(
-    task_id: int,
-    background: BackgroundTasks,
-    files: Optional[str] = None,
-) -> FileResponse:
-    """把 output 目录里的文件打包成 zip 一次性下载。"""
-    import tempfile
-    wanted = [n for n in files.split(",") if n] if files else None
-    if files is not None and not wanted:
-        raise HTTPException(400, "empty files list")
-    task, selected, partial = _select_task_output_files(task_id, wanted)
-    out_dir = _task_output_dir(task)
-    assert out_dir is not None  # _select_task_output_files 已经校验
-
-    tmp = tempfile.NamedTemporaryFile(suffix=".zip", delete=False)
-    tmp.close()
-    tmp_path = Path(tmp.name)
-    try:
-        _write_outputs_zip(tmp_path, out_dir, selected)
-    except Exception as e:
-        tmp_path.unlink(missing_ok=True)
-        bus.publish({
-            "type": "task_outputs_zip_failed",
-            "task_id": task_id,
-            "error": str(e),
-        })
-        raise
-
-    bus.publish({"type": "task_outputs_zip_ready", "task_id": task_id})
-    background.add_task(lambda: tmp_path.unlink(missing_ok=True))
-
-    basename = _task_archive_basename(task) or f"task_{task_id}"
-    archive_name = (
-        f"{basename}_outputs_selected.zip" if partial
-        else f"{basename}_outputs.zip"
-    )
-    return FileResponse(
-        tmp_path,
-        media_type="application/zip",
-        filename=archive_name,
-        background=background,
-    )
-
-
-@app.post("/api/queue/{task_id}/export-outputs")
-def export_task_outputs_to_data_exports(
-    task_id: int,
-    body: _ExportOutputsBody,
-) -> dict[str, Any]:
-    """把 output 文件打包保存到 data_exports/。"""
-    task, selected, partial = _select_task_output_files(task_id, body.files)
-    DATA_EXPORTS.mkdir(parents=True, exist_ok=True)
-    basename = _task_archive_basename(task) or f"task_{task_id}"
-    archive_name = f"{basename}_outputs_selected.zip" if partial else f"{basename}_outputs.zip"
-    dest = _unique_data_export_path(archive_name)
-    out_dir = _task_output_dir(task)
-    assert out_dir is not None
-    try:
-        _write_outputs_zip(dest, out_dir, selected)
-    except Exception as e:
-        dest.unlink(missing_ok=True)
-        bus.publish({"type": "task_outputs_zip_failed", "task_id": task_id, "error": str(e)})
-        raise
-    bus.publish({"type": "task_outputs_zip_ready", "task_id": task_id})
-    return _export_result(dest)
-
-
-@app.get("/api/queue/{task_id}/output/{filename:path}")
-def download_task_output(task_id: int, filename: str) -> FileResponse:
-    """下载 output 目录下的指定文件。"""
-    with db.connection_for() as conn:
-        task = db.get_task(conn, task_id)
-    if not task:
-        raise HTTPException(404)
-    out_dir = _task_output_dir(task)
-    if not out_dir or not out_dir.exists():
-        raise HTTPException(404, "no output dir")
-    f = _safe_output_relpath_or_400(out_dir, filename)
-    if not f.exists() or not f.is_file():
-        raise HTTPException(404, "file not found")
-    return FileResponse(
-        f,
-        media_type="application/octet-stream",
-        filename=f.name,
-    )
-
-
-@app.post("/api/queue/{task_id}/open-folder")
-def open_task_folder(task_id: int, request: Request) -> dict[str, Any]:
-    """在 server 主机上用 OS 文件管理器打开 output 目录。
-
-    **仅 loopback 请求允许**：云端部署时浏览器不在 server 那台机，开了用户也
-    看不到，反而是远程命令执行入口；这里直接 403 拒绝。
-    """
-    if not _is_loopback(request):
-        raise HTTPException(
-            403, "open-folder is only available for local (loopback) requests"
-        )
-    with db.connection_for() as conn:
-        task = db.get_task(conn, task_id)
-    if not task:
-        raise HTTPException(404)
-    out_dir = _task_output_dir(task)
-    if not out_dir or not out_dir.exists():
-        raise HTTPException(404, "no output dir")
-    try:
-        import platform
-        import subprocess as _sub
-        sysname = platform.system()
-        if sysname == "Windows":
-            os.startfile(str(out_dir))  # type: ignore[attr-defined]
-        elif sysname == "Darwin":
-            _sub.Popen(["open", str(out_dir)])
-        else:
-            _sub.Popen(["xdg-open", str(out_dir)])
-    except Exception as exc:
-        raise HTTPException(500, f"failed to open folder: {exc}") from exc
-    return {"opened": str(out_dir)}
-
-
-@app.delete("/api/queue/{task_id}")
-def delete_queue_item(task_id: int) -> dict[str, Any]:
-    with db.connection_for() as conn:
-        task = db.get_task(conn, task_id)
-        if not task:
-            raise HTTPException(404)
-        if task["status"] not in db.TERMINAL_STATUSES:
-            raise HTTPException(400, "only terminal tasks can be deleted")
-        db.delete_task(conn, task_id)
-    return {"deleted": task_id}
 
 
 # ---------------------------------------------------------------------------
